requirements.txt
wheel ~= 0.36.2
grpc-interceptor ~= 0.12.0
grpcio-tools ~= 1.30
flask ~= 1.1
Jinja2 ~= 2.11
pytest ~= 5.4